+++ src/app/GeneratorTeamsApp.ts
@@ -161,7 +161,9 @@ export class GeneratorTeamsApp extends Generator {
                     name: 'manifestVersion',
                     message: 'Which manifest version would you like to use?',
                     choices: versions,
-                    default: versions.find( (v: inquirer.objects.ChoiceOption) => v.extra.default)!.value,
+                    default: versions.find( (v: inquirer.objects.ChoiceOption) => v.extra.default) ? 
+                        versions.find( (v: inquirer.objects.ChoiceOption) => v.extra.default)!.value : 
+                        versions[0].value,
                     when: (answers: any) => (this.options.existingManifest && answers.updateManifestVersion && versions.length > 0) || (!this.options.existingManifest)
                 },
                 {